requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
pillow
python-multipart
ollama
python-dotenv
jinja2
